+++ code/system.ui/react.ui.monaco/src/common/const.mts
@@ -1,3 +1,5 @@
+import type { t } from '../common.t';
+
 export const COLORS = {
   BLACK: '#000',
   WHITE: '#fff',
@@ -11,4 +13,18 @@ export const CSS = {
   CLASS: {
     EDITOR: `sys-monaco-editor`,
   },
+} as const;
+
+export const LANGUAGES: t.EditorLanguage[] = [
+  'typescript',
+  'javascript',
+  'markdown',
+  'json',
+  'yaml',
+];
+
+export const DEFAULTS = {
+  tabSize: 2,
+  language: LANGUAGES[2],
+  className: CSS.CLASS.EDITOR,
 } as const;
