+++ code/system.ui/react.ui.monaco/src/ui/MonacoEditor/common.ts
@@ -1,2 +1,1 @@
 export * from '../common';
-export * from './common.const';
